Refactor (PlayerAnimationBP&PlayerCharacter) :
Major refactor of player controller, removed most cpp logic. For now all the Primary function does is activates a timer and sets bAttacking to true. In the Animation BP for the player the attack speed now dictates the animation speed of montage. There is also a HitNotify on Montages that sets Player bAttacking to false.